+++ src/tool/cppwinrt/cppwinrt/cppwinrt.vcxproj
@@ -23,7 +23,6 @@
     <ClInclude Include="..\strings\base_activation.h" />
     <ClInclude Include="..\strings\base_agile_ref.h" />
     <ClInclude Include="..\strings\base_array.h" />
-    <ClInclude Include="..\strings\base_async.h" />
     <ClInclude Include="..\strings\base_chrono.h" />
     <ClInclude Include="..\strings\base_collections.h" />
     <ClInclude Include="..\strings\base_collections_base.h" />
@@ -36,13 +35,10 @@
     <ClInclude Include="..\strings\base_collections_vector.h" />
     <ClInclude Include="..\strings\base_composable.h" />
     <ClInclude Include="..\strings\base_com_ptr.h" />
-    <ClInclude Include="..\strings\base_coroutine.h" />
-    <ClInclude Include="..\strings\base_coroutine_action.h" />
-    <ClInclude Include="..\strings\base_coroutine_action_with_progress.h" />
-    <ClInclude Include="..\strings\base_coroutine_fire_and_forget.h" />
-    <ClInclude Include="..\strings\base_coroutine_operation.h" />
-    <ClInclude Include="..\strings\base_coroutine_operation_with_progress.h" />
-    <ClInclude Include="..\strings\base_coroutine_resume.h" />
+    <ClInclude Include="..\strings\base_coroutine_foundation.h" />
+    <ClInclude Include="..\strings\base_coroutine_system.h" />
+    <ClInclude Include="..\strings\base_coroutine_threadpool.h" />
+    <ClInclude Include="..\strings\base_coroutine_ui_core.h" />
     <ClInclude Include="..\strings\base_delegate.h" />
     <ClInclude Include="..\strings\base_dependencies.h" />
     <ClInclude Include="..\strings\base_error.h" />
@@ -58,7 +54,6 @@
     <ClInclude Include="..\strings\base_meta.h" />
     <ClInclude Include="..\strings\base_natvis.h" />
     <ClInclude Include="..\strings\base_reference_produce.h" />
-    <ClInclude Include="..\strings\base_reflect.h" />
     <ClInclude Include="..\strings\base_security.h" />
     <ClInclude Include="..\strings\base_std_hash.h" />
     <ClInclude Include="..\strings\base_string.h" />
