--- src/tool/cppwinrt/strings/base_coroutine_action.h
@@ -1,16 +1,0 @@
-
-namespace std::experimental
-{
-    template <typename... Args>
-    struct coroutine_traits<winrt::Windows::Foundation::IAsyncAction, Args...>
-    {
-        struct promise_type final : winrt::impl::promise_base<promise_type,
-            winrt::Windows::Foundation::IAsyncAction,
-            winrt::Windows::Foundation::AsyncActionCompletedHandler>
-        {
-            void return_void() const noexcept
-            {
-            }
-        };
-    };
-}
